requirements.txt
certifi==2018.8.13
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-extensions==2.2.8
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
requests-mock==1.5.2
six==1.11.0
urllib3==1.23
whitenoise==4.0
